pyproject.toml
# [tool.poetry.dependencies]
# PyYAML = "^6.0"
# coloredlogs = "^15.0.1"
# docopt = "^0.6.2"
# python = "^3.10"
# selenium = "^4.10.0"
# speechrecognition = "^3.10.0"
# pydub = "^0.25.1"
# slack-sdk = "^3.21.3"
# pillow = "^10.2.0"


[project]
name = "mercari_bot"
version = "0.1.0"
description = "メルカリの出品アイテムを自動的に値下げするスクリプトです．"
authors = [
    { name = "KIMATA Tetsuya", email = "kimata@green-rabbit.net" }
]

readme = "README.md"
requires-python = ">= 3.10"

dependencies = [
    "docopt-ng>=0.9.0",
    "my-lib @ git+https://github.com/kimata/my-py-lib@336aa21adfe8a544d5e4170d73cda8f9f785fa32",
    "selenium>=4.31.0",
    "pydub>=0.25.1",
    "speechrecognition>=3.14.2",
    "pillow>=11.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sensing"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.rye]
managed = true
dev-dependencies = [
    "flaky>=3.8.1",
    "pytest-cov>=5.0.0",
    "pytest-freezegun>=0.4.2",
    "pytest-html>=4.1.1",
    "pytest-mock>=3.14.0",
    "pytest-playwright>=0.5.1",
    "pytest-xdist>=3.6.1",
    "pytest>=8.3.2",
]

[tool.pytest.ini_options]
minversion = "6.0"

# addopts = "--verbose --log-cli-level=DEBUG --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"

addopts = "--verbose --log-file-level=DEBUG --log-format=\"%(asctime)s %(levelname)s %(message)s\" --log-format=\"%(asctime)s %(levelname)s [%(filename)s:%(lineno)s %(funcName)s] %(message)s\" --capture=sys  --html=tests/evidence/index.htm --self-contained-html --cov=src --cov-report=html -vv"


testpaths = [
    "tests",
]
filterwarnings = [
    "ignore:The hookimpl CovPlugin.pytest_configure_node uses",
    "ignore:The hookimpl CovPlugin.pytest_testnodedown uses",
    "ignore::DeprecationWarning:pytest_freezegun",
]

[tool.coverage.run]
branch = true


[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",

  # 以下，未使用関数
  # webapp/log.py
  "if sqlite is not None:",
]

[tool.coverage.html]
directory = "tests/evidence/coverage"
